Pass depthFirst test
This (verbose) work passes the final test
for binarySearchTree.  It finds the deepest
node in the tree, stores the path to it, and
then performs a callback on each item in the
tree following that path, then the second
deepest path, then the third, etc.  Plenty of
refactoring and code cleanup is needed.